refactor(cache): split search index responsibilities
